+++ CHANGES.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Added documentation about supported types of fields for metadata (and info warnings for unsupported types). ([#329](https://github.com/lum-ai/odinson/pull/329))
 ### Changed
 - Use lucene's `KeywordAnalyzer` for parsed documents rather than `WhitespaceAnalyzer`.
+### Fixed
+- Negated lookaround queries return the correct spans (not off-by-one).
 
 ## [0.5.0] - 2021-08-07
 ### Added